Optimize asset retrieval and enhance test coverage
This commit introduces a more efficient method for asset retrieval in the LocalSolver by using the bufferManager. If the bufferManager already has directContractToAssets, the method will return it directly, resulting in performance optimization. Additionally, the TestManager now has the option to perform the datapraph flush. A new test case for handling a large amount of assets has also been added to the test suite. Finally, the BufferManager can now load assets directly from contracts, allowing to proload the contracts Asset factory for specific assets

+++ src/CsCannon/Tests/TestManager.php
@@ -28,17 +28,19 @@ class TestManager
     public  const ETHEREUM_TOKEN_ID = '47225'; //contract address
 
 
-    public static function initTestDatagraph(){
+    public static function initTestDatagraph($flushing = true){
 
 
 
 //         TestManagerPrivate::initTestDatagraph() ;
         //return
 
-        $sandraToFlush = new System('phpUnit_', true);
+        if ($flushing) {
+            $sandraToFlush = new System('phpUnit_', true);
 //        $system = new System('gossip', false, "127.0.0.1", "control_center");
-        \SandraCore\Setup::flushDatagraph($sandraToFlush);
-        $sandraToFlush->destroy();
+            \SandraCore\Setup::flushDatagraph($sandraToFlush);
+            $sandraToFlush->destroy();
+        }
         $system = new System('phpUnit_',true);
         $system->registerStructure = true ;
